se agrega delete y paginacion

+++ giovannilopez-chapur-backend/supplier/src/main/java/com/chapur/supplier/infrastructure/adapter/in/web/dto/PagedResponse.java
@@ -6,7 +6,7 @@ public record PagedResponse<T>(
         List<T> content,
         Integer page,
         Integer size,
-        Long totalElements,
+        Integer totalElements,
         Integer totalPage
 ) {
 }
